requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
docopt==0.6.2
idna==3.10
loguru==0.7.3
pipreqs==0.4.13
psutil==6.1.1
python-gnupg==0.5.4
requests==2.32.3
urllib3==2.3.0
yarg==0.1.10
ttkbootstrap==1.10.1